// enums.ts
export enum DirectionsEnum {
    HORIZONTAL = 'HORIZONTAL',
    VERTICAL = 'VERTICAL',
    DIAGONAL = 'DIAGONAL',
}

export enum OrientationsEnum {
    NORTH = 'NORTH',
    NORTHEAST = 'NORTHEAST',
    EAST = 'EAST',
    SOUTHEAST = 'SOUTHEAST',
    SOUTH = 'SOUTH',
    SOUTHWEST = 'SOUTHWEST',
    WEST = 'WEST',
    NORTHWEST = 'NORTHWEST'
}